pyproject.toml
[project]
name = "photo-watermark-helper"
version = "0.1.0"
description = "Advanced photo watermarking tool with server and CLI support"
authors = [
    {name = "colour93", email = "x@fur93.icu"},
]
dependencies = [
    "pillow>=11.3.0", 
    "piexif>=1.1.3", 
    "requests>=2.32.4", 
    "python-dotenv>=1.1.1", 
    "numpy>=2.3.1",
    "loguru>=0.7.2",
    "click>=8.1.7",
    "fastapi>=0.115.6",
    "uvicorn>=0.32.1",
    "rich>=13.9.4",
    "websockets>=14.1",
    "python-multipart>=0.0.19",
    "aiofiles>=24.1.0"
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
watermarker = "watermarker.main:cli"

[tool.pdm]
distribution = false
